+++ components/members/components/MemberDirectorySidebar/MemberPropertyItem.tsx
@@ -1,5 +1,4 @@
 import type { MemberPropertyType } from '@charmverse/core/prisma';
-import { useTheme } from '@emotion/react';
 import GithubIcon from '@mui/icons-material/GitHub';
 import InsertPhotoIcon from '@mui/icons-material/InsertPhoto';
 import LinkedInIcon from '@mui/icons-material/LinkedIn';
@@ -16,28 +15,6 @@ import { UpgradeChip } from 'components/settings/subscription/UpgradeWrapper';
 import { MEMBER_PROPERTY_CONFIG, PREMIUM_MEMBER_PROPERTIES } from 'lib/members/constants';
 import DiscordIcon from 'public/images/logos/discord_logo.svg';
 import FarcasterLogo from 'public/images/logos/farcaster.svg';
-
-function FarcasterIcon() {
-  const theme = useTheme();
-
-  if (theme.palette.mode === 'dark') {
-    return (
-      <SvgIcon viewBox='0 0 20 20' fontSize='small'>
-        <FarcasterLogo />
-      </SvgIcon>
-    );
-  } else {
-    return (
-      <img
-        style={{
-          width: 20,
-          height: 20
-        }}
-        src='/images/logos/farcaster_logo_grayscale.png'
-      />
-    );
-  }
-}
 
 export const MemberPropertyIcons: Record<MemberPropertyType, ReactNode> = {
   text: <TextIcon fontSize='small' />,
@@ -60,7 +37,11 @@ export const MemberPropertyIcons: Record<MemberPropertyType, ReactNode> = {
   google: <FaGoogle fontSize='small' />,
   telegram: <FaTelegramPlane fontSize='small' />,
   wallet: <FaWallet fontSize='small' />,
-  farcaster: <FarcasterIcon />
+  farcaster: (
+    <SvgIcon viewBox='0 0 20 20' fontSize='small'>
+      <FarcasterLogo />
+    </SvgIcon>
+  )
 };
 
 type Props = {
