+++ lib/members/interfaces.ts
@@ -55,7 +55,9 @@ export type Member = Pick<User, 'id' | 'createdAt' | 'updatedAt' | 'username'> &
   roles: Pick<Role, 'name' | 'id'>[];
   onboarded: boolean;
   searchValue: string;
-  farcasterUser?: FarcasterUser;
+  farcasterUser?: {
+    username: string;
+  };
 };
 
 export type CreateMemberPropertyPayload = Pick<MemberProperty, 'index' | 'name' | 'options' | 'type'>;
